requirements.txt
--only-binary :all: PyMuPDF==1.23.26
openai==1.14.0
python-dotenv==1.0.1
habanero==1.2.3
